Pull the `markDelivered` implementation up.
